pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "koten-text-refiner"
version = "0.1.0"
description = "Classical Japanese OCR error correction experiments with Qwen3.5-9B and Unsloth"
readme = "docs/qwen35_koten_ocr_plan.md"
requires-python = ">=3.11"
dependencies = [
  "datasets>=4.0.0",
  "evaluate>=0.4.0",
  "jiwer>=4.0.0",
  "mecab-python3>=1.0.10",
  "unidic-lite>=1.0.8",
  "orjson>=3.10.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "rapidfuzz>=3.9.0",
  "sacrebleu>=2.4.3",
  "scikit-learn>=1.5.0",
  "typer>=0.12.3",
  "transformers",
  "trl",
  "torch",
  "unsloth",
  "pytest>=8.3.0",
]

[project.scripts]
koten-refiner = "koten_refiner.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/koten_refiner"]

[tool.uv]
package = true
